# feat: Simplify is ins to an OR chain of eqs