--- a/src/daft-algebra/src/simplify/mod.rs
+++ b/src/daft-algebra/src/simplify/mod.rs
@@ -5,7 +5,7 @@ mod numeric;
 use boolean::{simplify_binary_compare, simplify_boolean_expr};
 use common_error::DaftResult;
 use common_treenode::{Transformed, TreeNode};
-use daft_dsl::{lit, Expr, ExprRef};
+use daft_dsl::{lit, Expr, ExprRef, LiteralValue};
 use daft_schema::schema::SchemaRef;
 use null::simplify_expr_with_null;
 use numeric::simplify_numeric_expr;
@@ -18,6 +18,7 @@ pub fn simplify_expr(expr: ExprRef, schema: &SchemaRef) -> DaftResult<Transforme
         simplify_expr_with_null,
         simplify_numeric_expr,
         simplify_misc_expr,
+        simplify_is_in_expr,
     ];
 
     // Our simplify rules currently require bottom-up traversal to work
@@ -52,14 +53,31 @@ fn simplify_misc_expr(expr: ExprRef, schema: &SchemaRef) -> DaftResult<Transform
                 .lt_eq(high.clone())
                 .and(between.clone().gt_eq(low.clone())),
         ),
-        // e IN () --> false
-        Expr::IsIn(_, list) if list.is_empty() => Transformed::yes(lit(false)),
         // CAST(e AS dtype) -> e if e.dtype == dtype
         Expr::Cast(e, dtype) if e.get_type(schema)? == *dtype => Transformed::yes(e.clone()),
         _ => Transformed::no(expr),
     })
 }
 
+const MAX_IS_IN_CHAIN_LENGTH: usize = 5;
+
+fn simplify_is_in_expr(expr: ExprRef, _schema: &SchemaRef) -> DaftResult<Transformed<ExprRef>> {
+    Ok(match expr.as_ref() {
+        Expr::IsIn(_, list) if list.is_empty() => Transformed::yes(lit(false)),
+        // If the list is a small literal list, we can just make it an OR chain of eqs, e.g.
+        // e IN (1, 2, 3) -> e = 1 OR e = 2 OR e = 3
+        Expr::IsIn(e, list)
+            if list.len() <= MAX_IS_IN_CHAIN_LENGTH
+                && list.iter().all(|e| matches!(e.as_ref(), Expr::Literal(l) if !matches!(l, LiteralValue::Series(_)))) =>
+        {
+            let chain_of_eqs = list
+                .iter()
+                .map(|item| e.clone().eq(item.clone()));
+            Transformed::yes(chain_of_eqs.reduce(|a, b| a.or(b)).unwrap())
+        }
+        _ => Transformed::no(expr),
+    })
+}
 #[cfg(test)]
 mod test {
     use std::sync::Arc;
@@ -159,6 +177,40 @@ mod test {
         let optimized = simplify_expr(input, &schema)?;
 
         assert!(optimized.transformed);
+        assert_eq!(optimized.data, expected);
+        Ok(())
+    }
+
+    #[rstest]
+    // One element list, can transform to eq
+    // e IN (1) --> e = 1
+    #[case(col("int").is_in(vec![lit(1)]), col("int").eq(lit(1)))]
+    // Small list, can transform to ORs
+    // e IN (1, 2, 3, 4, 5) --> e = 1 OR e = 2 OR e = 3 OR e = 4 OR e = 5
+    #[case(col("int").is_in(vec![lit(1), lit(2), lit(3), lit(4), lit(5)]), col("int").eq(lit(1)).or(col("int").eq(lit(2))).or(col("int").eq(lit(3))).or(col("int").eq(lit(4))).or(col("int").eq(lit(5))))]
+    fn test_is_in_exprs_can_chain_or_clauses(
+        #[case] input: ExprRef,
+        #[case] expected: ExprRef,
+        schema: SchemaRef,
+    ) -> DaftResult<()> {
+        let optimized = simplify_expr(input, &schema)?;
+
+        assert!(optimized.transformed);
+        assert_eq!(optimized.data, expected);
+        Ok(())
+    }
+
+    #[rstest]
+    // e IN (1, 2, 3, 4, 5, 6) --> e IN (1, 2, 3, 4, 5, 6)
+    #[case(col("int").is_in(vec![lit(1), lit(2), lit(3), lit(4), lit(5), lit(6)]), col("int").is_in(vec![lit(1), lit(2), lit(3), lit(4), lit(5), lit(6)]))]
+    fn test_is_in_exprs_more_than_max_chain_length(
+        #[case] input: ExprRef,
+        #[case] expected: ExprRef,
+        schema: SchemaRef,
+    ) -> DaftResult<()> {
+        let optimized = simplify_expr(input, &schema)?;
+
+        assert!(!optimized.transformed);
         assert_eq!(optimized.data, expected);
         Ok(())
     }